core/pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "kapybara"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Yanli 盐粒", email = "yanli@dify.ai"},
]
dependencies = ["pydantic>=2.12.5", "pydantic-settings>=2.12.0", "python-dotenv>=1.2.1", "logfire>=4.22.0", "rich>=14.3.2", "anyio>=4.12.1", "pydantic-ai-slim[openrouter]>=1.58.0", "uuid6>=2025.0.1", "tiktoken>=0.12.0"]
requires-python = "==3.14.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
    "ty>=0.0.15",
]



[tool.pdm.scripts]
lint = "ruff check . --fix"
format = "ruff format ."
fix = {composite = ["format", "lint"]}
typecheck = "ty check ."
test = { cmd = "pytest", env = { PYTHONPATH = "src" } }

[tool.ruff]
target-version = "py314"
extend-exclude = [".venv", ".pdm-build"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "PIE", "RUF"]
ignore = ["N818", "E501", "SIM108", "SIM103", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ty.src]
exclude = ["tests/"]
